pyproject.toml
[tool.poetry]
name = "differences_evaluator"
version = "0.1.0"
description = ""
authors = ["vcslav-v <vcslav.v@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3.1"

[tool.poetry.scripts]
gendiff = "differences_evaluator.scripts.gendiff:main"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest-cov = "^2.10.1"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
